+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Refactored nostr utility methods.
 - Fixed numerous issues regarding subscription handling.
+- Updated subscriptions config page.
 
 ## [0.9.0]